--- packages/node-sdk/sm/primitives/evm/rpc/dynamic-primitive.ts
@@ -1,94 +1,0 @@
-import { registerDynamicPrimitive } from "@paima/db";
-import { World } from "@paima/coroutine";
-import type { StateUpdateStream } from "@paima/coroutine";
-import type {
-  ConfigPrimitivePayloadType,
-  ConfigSyncProtocolType,
-  FlattenSyncProtocolIOFor,
-  TPrimitiveErc721Config,
-  TPrimitiveEvmGenericConfig,
-} from "@paima/config";
-import {
-  type DeepMutable,
-  type EvmAddress,
-  generateDynamicPrimitivePrefix,
-} from "@paima/utils";
-import {
-  ConfigPrimitiveType,
-  GlobalConfig,
-  networkConfigQuery,
-} from "@paima/config";
-import assertNever from "assert-never";
-import { hashConfig } from "@paima/runtime";
-
-/**
- * We omit storing the name from the config
- * because it's dynamically generated on the db insert
- */
-type StoredConfig<T> = Omit<T, "displayName">;
-
-export default function* processDynamicEvmPrimitiveSyncProtocolResponse(
-  response: FlattenSyncProtocolIOFor<
-    ConfigSyncProtocolType.EVM_RPC_PARALLEL,
-    ConfigPrimitiveType.EvmRpcDynamicPrimitive,
-    ConfigPrimitivePayloadType.Event
-  >,
-): StateUpdateStream<void> {
-  let type:
-    | ConfigPrimitiveType.EvmRpcERC721
-    | ConfigPrimitiveType.EvmRpcGeneric;
-  let config:
-    | StoredConfig<TPrimitiveEvmGenericConfig>
-    | StoredConfig<TPrimitiveErc721Config>;
-
-  // the numeric id is generated by the insert query.
-  const baseName = generateDynamicPrimitivePrefix(
-    response.output.syncProtocol.payload.primitiveName,
-  );
-
-  const networkQuery = networkConfigQuery(GlobalConfig.networks);
-
-  const networkConfig = networkQuery.networkForCaip2(
-    response.output.syncProtocol.payload.caip2,
-  );
-  if (networkConfig == null) {
-    throw new Error(
-      `No network found in config for caip2: ${response.output.syncProtocol.payload.caip2}`,
-    );
-  }
-  switch (response.output.payload.targetConfig.type) {
-    case ConfigPrimitiveType.ERC721:
-      type = ConfigPrimitiveType.ERC721;
-      config = {
-        ...response.output.payload.targetConfig,
-        contractAddress: response.output.payload.contractAddress
-          .toLowerCase() as EvmAddress,
-        startBlockHeight:
-          response.output.syncProtocol.payload.ownChain.blockNumber,
-      };
-      break;
-
-    case ConfigPrimitiveType.EvmGeneric:
-      type = ConfigPrimitiveType.EvmGeneric;
-      config = {
-        ...response.output.payload.targetConfig,
-        contractAddress: response.output.payload.contractAddress
-          .toLowerCase() as EvmAddress,
-        startBlockHeight:
-          response.output.syncProtocol.payload.ownChain.blockNumber,
-      };
-      break;
-    default:
-      assertNever.default(response.output.payload.targetConfig);
-  }
-
-  yield* World.resolve(registerDynamicPrimitive, {
-    base_name: baseName,
-    config: structuredClone(config) as DeepMutable<typeof config>,
-    config_hash: hashConfig(config),
-    syncProtocol_name: response.output.syncProtocol.name,
-    parent_name: response.output.syncProtocol.payload.primitiveName,
-    primitive_caip2: response.output.syncProtocol.payload.caip2,
-    primitive_type: response.output.payload.targetConfig.type,
-  });
-}
