+++ packages/paima-sdk/concise/src/v2/builtins/index.ts
@@ -1,3 +1,3 @@
 export * from "./batcher.ts";
 export * from "./grammar.ts";
-export * from "./primitives.ts";
+// export * from "./primitives.ts";
