+++ packages/paima-sdk/config/src/schema/primitive/config/types.ts
@@ -6,7 +6,7 @@ import type { PrimitiveConfig } from "../../../config/parts/primitive.ts";
 export enum ConfigPrimitiveType {
   EvmRpcGeneric = "evm-rpc-generic",
   EvmRpcPaimaL2 = "evm-rpc-paima-l2",
-  EvmRpcERC20 = "evm-rpc-erc20",
+  // EvmRpcERC20 = "evm-rpc-erc20",
   EvmRpcERC721 = "evm-rpc-erc721",
   EvmRpcERC6551Registry = "evm-rpc-erc6551-registry",
   EvmRpcERC1155 = "evm-rpc-erc1155",
@@ -61,9 +61,9 @@ export const EvmPrimitivesToSyncProtocol = {
   [ConfigPrimitiveType.EvmRpcPaimaL2]: [
     ConfigSyncProtocolType.EVM_RPC_PARALLEL,
   ],
-  [ConfigPrimitiveType.EvmRpcERC20]: [
-    ConfigSyncProtocolType.EVM_RPC_PARALLEL,
-  ],
+  // [ConfigPrimitiveType.EvmRpcERC20]: [
+  //   ConfigSyncProtocolType.EVM_RPC_PARALLEL,
+  // ],
   [ConfigPrimitiveType.EvmRpcERC721]: [
     ConfigSyncProtocolType.EVM_RPC_PARALLEL,
   ],
